better code for deep denoising
